+++ src/Ext/Techno/Body.cpp
@@ -1036,7 +1036,7 @@ void TechnoExt::ExtData::Serialize(T& Stm)
 		.Process(this->ForceFullRearmDelay)
 		.Process(this->DelayedFire_Anim)
 		.Process(this->DelayedFire_Anim_LoopCount)
-		.Process(this->DelayedFire_DurationTimer)
+		.Process(this->DelayedFire_Duration)
 		;
 }